% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bane.R
\docType{class}
\name{Bane}
\alias{Bane}
\title{Bane Class}
\description{
R6 class to implement the BaNE model
}
\details{
Bane class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{indep}}{independent variable names}

\item{\code{dep}}{list of variables dependencies}

\item{\code{mu}}{independent variable priors}

\item{\code{lm}}{dependent variable priors}

\item{\code{n}}{Number of independent nodes}

\item{\code{m}}{Number of dependent nodes}

\item{\code{k}}{Number of total nodes}

\item{\code{N}}{Rows of data}

\item{\code{data}}{Data frame containing phenotype data}

\item{\code{all_conds}}{Character vector of all condtion names}

\item{\code{param_names}}{Names of all model parameters}

\item{\code{J}}{Total number of parameters}

\item{\code{M}}{Structure of the parameter matrix}

\item{\code{M_}}{A helper object}

\item{\code{q_}}{A helper object}

\item{\code{mu_pr}}{The mean vector for the multivariate normal prior}

\item{\code{Sig_inv}}{The precision matrix for the multivariate normal prior}

\item{\code{LD_Data}}{Data formatted as a list for LaplacesDemon}

\item{\code{ld}}{Output from LaplacesDemon, the posterior sampler}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Bane$new()}}
\item \href{#method-params_to_mat}{\code{Bane$params_to_mat()}}
\item \href{#method-LLX}{\code{Bane$LLX()}}
\item \href{#method-LPR}{\code{Bane$LPR()}}
\item \href{#method-LD_Model}{\code{Bane$LD_Model()}}
\item \href{#method-run_chain}{\code{Bane$run_chain()}}
\item \href{#method-maximum_likelihood}{\code{Bane$maximum_likelihood()}}
\item \href{#method-post_subs}{\code{Bane$post_subs()}}
\item \href{#method-plot_topology}{\code{Bane$plot_topology()}}
\item \href{#method-clone}{\code{Bane$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new BaNE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$new(indep, dep, mu, lm, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{indep}}{A character vector of independent variable names.}

\item{\code{dep}}{A named list of dependent variables with their dependencies. List names are the dependent variable names and list elements are the corresponding vector of dependency names.}

\item{\code{mu}}{Prior proportions for independent variables.}

\item{\code{lm}}{Prior baseline proportions for dependent variables.}

\item{\code{data}}{A dataframe whose columns are referenced by dep and indep}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Bane` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-params_to_mat"></a>}}
\if{latex}{\out{\hypertarget{method-params_to_mat}{}}}
\subsection{Method \code{params_to_mat()}}{
Helper function to structure parameters as a matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$params_to_mat(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Vector of model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size (k+1) x k
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LLX"></a>}}
\if{latex}{\out{\hypertarget{method-LLX}{}}}
\subsection{Method \code{LLX()}}{
Log likelihood
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$LLX(p, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Vector of model parameters}

\item{\code{data}}{A data matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log likelihood value of `p` given `data`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LPR"></a>}}
\if{latex}{\out{\hypertarget{method-LPR}{}}}
\subsection{Method \code{LPR()}}{
Log prior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$LPR(p)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Vector of model parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log prior evaluated at `p`
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LD_Model"></a>}}
\if{latex}{\out{\hypertarget{method-LD_Model}{}}}
\subsection{Method \code{LD_Model()}}{
Log posterior
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$LD_Model(p, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{Vector of model parameters}

\item{\code{data}}{A data matrix}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A posterior evalutions in LaplacesDemon format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_chain"></a>}}
\if{latex}{\out{\hypertarget{method-run_chain}{}}}
\subsection{Method \code{run_chain()}}{
Run the HMC sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$run_chain(Iterations = 2000, Status = 100, Thinning = 5, eps = 0.1, L = 3)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Iterations}}{passed to `LaplacesDemon`}

\item{\code{Status}}{passed to `LaplacesDemon`}

\item{\code{Thinning}}{passed to `LaplacesDemon`}

\item{\code{eps}}{passed to `LaplacesDemon`}

\item{\code{L}}{passed to `LaplacesDemon`}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A LaplacesDemon object contain sampler info and posterior draws
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-maximum_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-maximum_likelihood}{}}}
\subsection{Method \code{maximum_likelihood()}}{
Mode/ maximum likelihood
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$maximum_likelihood()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Output from the `optim` function maximizng the likelihood function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_subs"></a>}}
\if{latex}{\out{\hypertarget{method-post_subs}{}}}
\subsection{Method \code{post_subs()}}{
Posterior sampled subcohort proportions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$post_subs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with subcohort, corresponding proportion draws, and a plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_topology"></a>}}
\if{latex}{\out{\hypertarget{method-plot_topology}{}}}
\subsection{Method \code{plot_topology()}}{
Plot the phenotype topology
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$plot_topology()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot with phenotype topology
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bane$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
