% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bane_example.R
\name{bane_example}
\alias{bane_example}
\title{BaNE examples}
\usage{
bane_example(example = "two_conditions")
}
\arguments{
\item{example}{One of "all_conditions" or "two_conditions". View source code for model specifications.}
}
\value{
A plot of posterior subcohort proportions
}
\description{
Run one of a few examples using a built-in dataset
}
