% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bane_shiny.R
\name{bane_shiny}
\alias{bane_shiny}
\title{BaNE Shiny App}
\usage{
bane_shiny(data)
}
\arguments{
\item{data}{A data frame to use the model with}
}
\value{
nothing, usually
}
\description{
Runs a GUI for easily implementing the BaNE model
}
