% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_normal.R
\name{estimate_normal}
\alias{estimate_normal}
\title{Estimate Normal Distribution of Inverse Sigmoid Transformed Beta}
\usage{
estimate_normal(A, B, size = 1e+05)
}
\arguments{
\item{A}{The alpha shape parameter}

\item{B}{The beta shape parameter}

\item{size}{The sample size}
}
\value{
a list of values of interest
}
\description{
Given beta shape parameters A and B, this will estimate the mean
and variance of the inverser sigmoid transformed distribution via
a normal distribution.
}
\details{
This is useful for choosing prior values. Prior values should be given in
terms of log odds, but sometimes proportions are more intuitively appealing.
This function will estimate the log odds distibution given the corresponding
beta distribution of a proprtion.
}
